Fail soft where possible. New -q option.
New --fail-on-errors option to die immediately on errors, default is to
try to carry on. This is an acknowledgement of the reality that our
file transfers are slow, and sometimes the source changes from under
us. Don't fail an overnight job that would have 98% succeeded just
because of an HTTP failure.

Call out filesystem-changing options and major operations by default,
suppress with -q.