Save a reference to the dest FileHash.
If we download a file and/or change its metadata, the metadata for the
file as it used to be is still stored in the local HSYNC.SIG. This can
result in a lot of unnecessary rehashing.

In hashlist_check(), save a reference to the object referring to the
local file. Then, when we change the file, we can update the metadata to
reflect the new contents. (Not done the update yet, just store the ref.)