New debug options for scan and dir check.
Add --scan-debug and --check-debug to dump the file lists at
hashlist_generate() and hashlist_check() time. Write some test code to
fetch this debug output and give us a chance to compare the lists.

This will give us a chance to functionally test inclusion and exclusion
code in a relatively sensible way.