Fix horror bug in hashlist_generate().
We get the list dirs from os.walk(). We need to make changes to that
list to prune the filesystem walk. This is all well and good, until
someone deletes items from the list over which we're iterating... That
breaks things. Directories disappear from the list apparently at random,
though actually it's quite predictable.

A better way to do this exists, but the way in this commit works.