Refactor HashList, allow memory /or/ Sqlite
Turns out the performance penalty of sqlite3 HashList is very high. So,
subclass HashList, and make the default an implementation using builtin
types (but following HashList semantics), and SqliteHashList a subclass
that goes to the disk.

This has a few nice side-effects. First, a HashList gets more checks
than a list in RAM, and also has proper tests. It is a bit slower, but
is unlikely to be a dealbreaker.

Plus, the user is in control. If they care more about the memory hit,
they can tell the program to use the slow but smaller option.