Use OrderedDict to pre-sort the HashDict
The big memory drain, even in the Sqlite-backed code, was the sort of the
HashDict in hashlist_check(), which generates the list of files to be
fetched or deleted. It pulls the whole list into ram, which defeats the
point of the disk backend.

Turns out there's no need to do that. If the HashDict uses an
OrderedDict and the items are inserted into it in sorted order, a simple
iteritems() is enough, allowing the generator to pull the items from the
database piecemeal.