Try to guess the other side's sigfile name
This is a common sig/lock detection case. The source overrides the
HSYNC.SIG* name for some reason, e.g. ubuntu-HSYNC.SIG. The dest does
not. For some reason the lockfile ends up in the sigfile, and the remote
attempts to download it (and fails).

Of course, it shouldn't be in the signature file either. Hopefully all
our paranoia has fixed that.